GRHB-42: * helpers and components files moved also changed types import source
