+++ frontend/src/components/uam/common/enums/user-table-accesson.enum.ts
@@ -1,0 +1,6 @@
+enum UserTableAccesson {
+  ID = 'id',
+  EMAIL = 'email',
+}
+
+export { UserTableAccesson };
