+++ frontend/src/components/uam/components/components.ts
@@ -1,0 +1,1 @@
+export { UsersTable } from './users-table/users-table';
