+++ frontend/src/components/uam/helpers/get-rows-helper/get-rows.helper.ts
@@ -1,0 +1,16 @@
+import {
+  UsersGetAllItemResponseDto,
+  UsersGetAllResponseDto,
+} from 'common/types/types';
+import { UserTableAccesson } from 'components/uam/common/enums/user-table-accesson.enum';
+
+const getRows = (
+  users: UsersGetAllResponseDto,
+): UsersGetAllItemResponseDto[] => {
+  return users.map((user) => ({
+    [UserTableAccesson.ID]: user.id,
+    [UserTableAccesson.EMAIL]: user.email,
+  }));
+};
+
+export { getRows };
