+++ frontend/src/components/uam/helpers/helpers.ts
@@ -1,0 +1,2 @@
+export { getColumns } from './get-columns-helper/get-columns.helper';
+export { getRows } from './get-rows-helper/get-rows.helper';
